can: replace timestamp as unique skb attribute
Commit 514ac99c64b "can: fix multiple delivery of a single CAN frame for
overlapping CAN filters" requires the skb->tstamp to be set to check for
identical CAN skbs.

Without timestamping to be required by user space applications this timestamp
was not generated which lead to commit 36c01245eb8 "can: fix loss of CAN
frames in raw_rcv" - which forces the timestamp to be set in all CAN related
skbuffs by introducing several __net_timestamp() calls.

This forces e.g. out of tree drivers to add __net_timestamp() after skbuff
creation to prevent the frame loss fixed in mainline Linux.

This patch removes the timestamp dependency and uses an atomic counter to
create an unique identifier together with the skbuff pointer.

Btw. the new skbcnt element introduced in struct can_skb_priv has to be
initialized with zero in out-of-tree drivers too.

Signed-off-by: Oliver Hartkopp <socketcan@hartkopp.net>
Cc: linux-stable <stable@vger.kernel.org>
Signed-off-by: Marc Kleine-Budde <mkl@pengutronix.de>